requirements.txt
PyQt5==5.15.4
qrcode==7.3.1
Pillow==9.4.0
